Refactoring and DRYing up
